Fix an issue with large inputs to hashers in 32-bit executables
One of our SHA-384 unit tests builds a 1GiB string, which means 8
billions of bits. That test failed in 32-bit mode, despite
update_count() now taking a long long.

+++ breeze/cryptography/brz/merkle_damgard_machine.tpp
@@ -127,7 +127,7 @@ merkle_damgard_machine< Engine >::do_append( RandomIter begin,
 
     // bufferize any remaining input; update the bit count
     std::copy( curr, end, breeze::begin( m_input_buffer ) + index ) ;
-    update_count( byte_width * ( end - begin ) ) ;
+    update_count( byte_width * static_cast< long long >( end - begin ) ) ;
 }
 
 template< typename Engine >
